refine PlotMultiBar demos
